refactor(typos): Simplify the top-level API
